Some tiny modifications (#81)
* Update 01-basics-01-bayes-rule.Rmd

delete the duplicated example and fix a typo: the an -> an

* Update 01-basics-03-freq-vs-bayes.Rmd

different colors misled me to interpret the sample as the sample space. Change sample to one sample

* Update 01-basics-03-freq-vs-bayes.Rmd

different colors misled me to interpret the sample as the sample space. Change sample to one sample

* In an answer professor Christian Robert corrected that statement, please see the answer. 

https://stats.stackexchange.com/a/508294/103153

* Error fixed

Error in typing which would lead to misunderstanding